Add generic component update signal mixin
